Don't force rich text formatting on app, summary and body labels.
QLabel will not propagate mouseReleaseEvent when clicking on text when
rich text is enabled. This means that clicking the notification does
nothing while the indended behaviour is to execute the default action.

This workaround disables rich text by default, which makes mouseReleaseEvent
work again. Rich text can still be used as QLabel uses AutoText by default.

This also means that if a text contains rich text, then clicking on it will
not have the expected behaviour, but it's still better than never having
the intended behaviour !