Use QLatin1String where it is more efficient
QLatin1String, which is just a thin wrapper around char * that specify the
encoding. There are overloads taking QLatin1String for functions that can
operate on the raw latin1 data directly without conversion.

Avoid having multiple QStringLiterals with the same content. For
QLatin1String, compilers try to consolidate identical literals so that they
are not duplicated. For QStringLiteral, identical strings cannot be merged.