Never miss a notification
Closes https://github.com/lxqt/lxqt/issues/1573

The list is on a menu that can be shown by left/right clicking the new notification tray icon. The icon appears only if there is an unattended notification and, of course, if its spin-box has a positive value in `Desktop Notifications` → `Advanced Settings`. There is also an option for blacklisting.

If a menu-item is clicked, its corresponding notification will be shown and considered attanded, so that it'll be removed from the list. The actuall list is saved in `~/.cache/lxqt-notificationd/unattended.list` (thanks to @Vladimir-csp). Etc.