Added a preview button, instead of automatic preview
With it, it's less likely that the preview notification starts at the previous position. With automatic preview, it always started there and then went to the new position (because there's no way to know exactly when the config file is ready).

The preview button can be clicked by pressing `Enter` too.

Closes https://github.com/lxqt/lxqt-notificationd/issues/236