No focused widget inside notifications
This patch just sets the focus policy of buttons and combo-boxes to `Qt::NoFocus` inside the notification window.

Under X11, it makes no difference, because the notification window is never activated (we set `Qt::X11BypassWindowManagerHint`).

But, under Wayland, where window activation is controlled by the WM, it avoids troubles and even prevents crashes, that are caused by bugs in Qt.

Closes https://github.com/lxqt/lxqt-notificationd/issues/329